requirements.txt
Django
djangorestframework
django-crispy-forms
crispy-bootstrap4
lxml
Pillow
concurrent-log-handler
apscheduler
markdown
inflect
django-environ
google-genai
pytz
django-cleanup
python-toon
cryptography
openai
anthropic
